requirements.txt
flask
flask-cors
gunicorn
pandas
numpy
scikit-learn
scipy
requests
python-dotenv
